+++ packages/blockera/php/Providers/EditorAssetsProvider.php
@@ -3,6 +3,7 @@
 namespace Blockera\Setup\Providers;
 
 use Blockera\Setup\Blockera;
+use Blockera\Telemetry\Config;
 use Blockera\WordPress\RenderBlock\Setup;
 use Illuminate\Contracts\Container\BindingResolutionException;
 
@@ -121,7 +122,22 @@ class EditorAssetsProvider extends \Blockera\Bootstrap\AssetsProvider {
 			return $inline_script;
 		}
 
-		return sprintf( '%s%s', $inline_script . PHP_EOL, $script );
+		$telemetry_debug_data = [
+			'php_version' => phpversion(),
+			'wordpress_version' => get_bloginfo('version'),
+			'product_slug' => Config::getConsumerConfig('name'),
+			'product_version' => Config::getConsumerConfig('version'),
+		];
+
+		return sprintf(
+			'%s%s
+			window.blockeraOptInStatus = "' . get_option( blockera_core_config( 'telemetry.options.opt_in_status' ), null ) . '";
+			window.blockeraCommunityUrl = "' . blockera_core_config( 'telemetry.community_url' ) . '";
+			window.blockeraPermissionsLink = "' . blockera_core_config( 'telemetry.permissions_link' ) . '";
+			window.blockeraTelemetryDebugData = ' . wp_json_encode( $telemetry_debug_data ) . ';',
+			$inline_script . PHP_EOL,
+			$script
+		);
 	}
 
 	/**